Move Blockchain to test_utils module
